Changes to data structure
based on citation chasing, I found that the basic structure of the graph
should be:

lp-index: a map between Hashcode (Integer) and the physical index
(Integer)
pl-index: an Integer array with indices correlating to the lp-index
values and elements corresponding to the logical ID of the neighborhoods
adjacency index: an array of VertexRecord objects. The indices of the
array correlate with the pl-index indices exactly.