update to testDeleteEdgeTwice
expected behavior is to throw an exception since deleteEdge does not
have a return value to indicate an edge was not removed. Perhaps this
should have been a boolean, but I'm time limited at this point and can't
reasonably update the method and all interdependencies in the DS and
tests.